Extract DecodingConfig
